Refactored forms datastore, added forms file and common store.
